[refactor] move applocation.ex into situation_room.ex app definition
